requirements.txt
langchain==0.2.0
langchain-openai==0.1.0
langchain-community==0.2.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
pypdf==4.1.0
python-dotenv==1.0.1
openai==1.35.0

